refactor(login): move utils/* and hooks/* in Login/ page to global
Move all utility functions and hooks in Login page to global
level directory, e.g., src/{utils,hooks}/ to share logic with Register
page in the future
