+++ src/utils/validateFormData.js
@@ -2,7 +2,11 @@ import { z } from 'zod';
 
 const errorMessage = {
   email: 'ที่อยู่อีเมล์ไม่ถูกต้อง กรุณาลองใหม่อีกครั้ง',
-  password: 'รหัสผ่านไม่ถูกต้อง กรุณาลองใหม่อีกครั้ง',
+  password: {
+    string: 'รหัสผ่านไม่ถูกต้อง กรุณาลองใหม่อีกครั้ง',
+    min: 'รหัสผ่านต้องอย่างน้อย 8 ตัวอักษร',
+    max: 'รหัสผ่านต้องไม่เกิน 24 ตัวอักษร',
+  },
 };
 
 const loginFormSchema = z.object({
@@ -13,10 +17,10 @@ const loginFormSchema = z.object({
     .email({ message: errorMessage.email }),
   password: z
     .string({
-      message: errorMessage.password,
+      message: errorMessage.password.string,
     })
-    .min(8, errorMessage.password)
-    .max(24, errorMessage.password),
+    .min(8, errorMessage.password.min)
+    .max(24, errorMessage.password.max),
 });
 
 /**
